ordenando metas

+++ src/http/routes/create-goal.ts
@@ -9,7 +9,7 @@ export const createGoalRoute: FastifyPluginAsyncZod = async (app) => {
             schema: {
                 body: z.object({
                     title: z.string(),
-                    desiredWeeklyFrequency: z.number().int().min(1).max(7),
+                    desiredWeeklyFrequency: z.number().min(1).max(7),
                 }),
             },
         },
